# Faster integer formatting