This PR provides an alternative implementation for a long → decimal conversion.

The main difference is that it uses an algorithm pulled from https://github.com/jeaiii/itoa. The source there is C++, it was converted by hand to C for inclusion with this gem. A writeup of this algorithm can be found [here](https://jk-jeon.github.io/posts/2022/02/jeaiii-algorithm/). jeaiii's algorithm is covered by the MIT License, see source code.

On addition this version now also generates the string directly into the `fbuffer`, foregoing the need to run a separate memory copy.

As a result, I see a speedup of 32% on Apple Silicon M1 for an integer set of benchmarks. 

```
== Encoding ints (45025 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
        json (local)     1.566k i/100ms
Calculating -------------------------------------
        json (local)     16.322k (± 3.9%) i/s   (61.27 μs/i) -     82.998k in   5.093499s

Normalize to 8009 byte
== Encoding ints (45025 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
       json (2.10.2)     1.228k i/100ms
Calculating -------------------------------------
       json (2.10.2)     12.352k (± 0.5%) i/s   (80.96 μs/i) -     62.628k in   5.070472s
```

These benchmarks are run via a script ([link](https://gist.github.com/radiospiel/04019402726a28b31616df3d0c17bd1c)) which is based on the gem's `benchmark/encoder.rb` file. There are probably better ways to run benchmarks :) My version allows to combine multiple test cases into a single one. 

The script adds a "cents" test case ([link](https://gist.github.com/radiospiel/04019402726a28b31616df3d0c17bd1c#file-encoder-rb-L188)) which explicitly covers integers in a range that is relevant for financial institutions. (Comparing benchmarks for that case show a speedup of 41%)

The `dumps` benchmark, which covers the JSON files in `benchmark/data/*.json`, sees a speedup of ~8 %.