@radiospiel 2025-03-17
Note: see [this](https://github.com/ruby/json/pull/768#issuecomment-2728622581) for discussing licensing concerns. Marking this PR as draft for the time being.