@radiospiel 2025-03-21
The windows build errors have been fixed. This PR is ready for review.

The changes are most impactful for long integers. A tests on 19 digit numbers shows a 65% speedup; the "cents" test (a mix of numbers between 1 and 7 digits) has 41%. On 2-digit numbers the gains are still  ~11%. (all tested on macOS, Apple Silicon M1)

I think a 32-bit version could be built and might over some small gains. The C++ version, which is the base for these changes, specialises templates for various integer widths. I didn't want to go /that/ far here.